Refactored ResourceList-related code in Razor Pages web app project to use shared service (instead of db context) directly, and to prepare it for new Unit Tests
