+++ src/NetLearner.Pages.Tests/ResourceListPageTests.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using NetLearner.Pages.ResourceLists;
+using System.Threading.Tasks;
+
+namespace NetLearner.Pages.Tests
+{
+    public class ResourceListPageTests
+    {
+        public async Task OnGetAsync_LoadsResourceLists()
+        {
+            // Arrange
+            //var pageModel = new IndexModel(mockAppDbContext.Object);
+        }
+    }
+}
