+++ src/NetLearner.Pages.Tests/UnitTest1.cs
@@ -1,0 +1,14 @@
+using System;
+using Xunit;
+
+namespace NetLearner.Pages.Tests
+{
+    public class UnitTest1
+    {
+        [Fact]
+        public void Test1()
+        {
+            Assert.True(true);
+        }
+    }
+}
