+++ src/NetLearner.Pages/Pages/ResourceLists/Create.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model NetLearner.Pages.CreateModel
+@model NetLearner.Pages.ResourceLists.CreateModel
 
 @{
     ViewData["Title"] = "Create";
