+++ src/NetLearner.Pages/Pages/ResourceLists/Create.cshtml.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using NetLearner.SharedLib.Data;
 using NetLearner.SharedLib.Models;
 
-namespace NetLearner.Pages
+namespace NetLearner.Pages.ResourceLists
 {
     public class CreateModel : PageModel
     {
