+++ src/NetLearner.Pages/Pages/ResourceLists/Delete.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model NetLearner.Pages.DeleteModel
+@model NetLearner.Pages.ResourceLists.DeleteModel
 
 @{
     ViewData["Title"] = "Delete";
