+++ src/NetLearner.Pages/Pages/ResourceLists/Details.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model NetLearner.Pages.DetailsModel
+@model NetLearner.Pages.ResourceLists.DetailsModel
 
 @{
     ViewData["Title"] = "Details";
