+++ src/NetLearner.Pages/Pages/ResourceLists/Edit.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model NetLearner.Pages.EditModel
+@model NetLearner.Pages.ResourceLists.EditModel
 
 @{
     ViewData["Title"] = "Edit";
