+++ src/NetLearner.Pages/Pages/ResourceLists/Edit.cshtml.cs
@@ -9,7 +9,7 @@ using Microsoft.EntityFrameworkCore;
 using NetLearner.SharedLib.Data;
 using NetLearner.SharedLib.Models;
 
-namespace NetLearner.Pages
+namespace NetLearner.Pages.ResourceLists
 {
     public class EditModel : PageModel
     {
