+++ src/NetLearnerApp.sln
@@ -21,7 +21,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{B649C5B1
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetLearner.SharedLib.Tests", "NetLearner.SharedLib.Tests\NetLearner.SharedLib.Tests.csproj", "{494458B4-B85A-42B6-BFCC-242247C01CD7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetLearner.Mvc.Tests", "NetLearner.Mvc.Tests\NetLearner.Mvc.Tests.csproj", "{1EA6E097-BE1D-4F28-A720-22A0D3785340}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetLearner.Mvc.Tests", "NetLearner.Mvc.Tests\NetLearner.Mvc.Tests.csproj", "{1EA6E097-BE1D-4F28-A720-22A0D3785340}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetLearner.Pages.Tests", "NetLearner.Pages.Tests\NetLearner.Pages.Tests.csproj", "{86A29A92-73D9-405E-95C0-886E0C20896B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -65,6 +67,10 @@ Global
 		{1EA6E097-BE1D-4F28-A720-22A0D3785340}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1EA6E097-BE1D-4F28-A720-22A0D3785340}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1EA6E097-BE1D-4F28-A720-22A0D3785340}.Release|Any CPU.Build.0 = Release|Any CPU
+		{86A29A92-73D9-405E-95C0-886E0C20896B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{86A29A92-73D9-405E-95C0-886E0C20896B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{86A29A92-73D9-405E-95C0-886E0C20896B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{86A29A92-73D9-405E-95C0-886E0C20896B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -72,6 +78,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{494458B4-B85A-42B6-BFCC-242247C01CD7} = {B649C5B1-571D-4953-865F-51CA8728156B}
 		{1EA6E097-BE1D-4F28-A720-22A0D3785340} = {B649C5B1-571D-4953-865F-51CA8728156B}
+		{86A29A92-73D9-405E-95C0-886E0C20896B} = {B649C5B1-571D-4953-865F-51CA8728156B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1630246B-29C2-4F69-830E-872F1F40E28F}
